src/config.py
"""
配置模块 - 读取和加载用户配置文件
"""

import yaml
from typing import Any, Dict


class Config:
    """加载和管理配置文件"""

    def __init__(self, config_path: str = "config.yaml"):
        self.config_path = config_path
        self._data: Dict[str, Any] = {}

    def load(self) -> bool:
        """加载 YAML 配置文件"""
        pass

    def get(self, key: str, default=None) -> Any:
        """获取配置项，支持点号分隔的嵌套访问"""
        pass

    def validate(self) -> list[str]:
        """验证配置完整性，返回缺失项列表"""
        pass
