requirements.txt
requests>=2.31.0
pyyaml>=6.0
opencv-python>=4.8.0
mss>=9.0.0
pydirectinput>=1.0.4
pynput>=1.7.6
pywin32>=306
